Move Model objects to gameObjects.Model namespace
