+++ GameObjects/ControlPanel.cs
@@ -4,6 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Windows.Input;
+using Action = GameObjects.Model.Action;
 
 namespace GameObjects
 {
