+++ GameObjects/GameObjects.csproj
@@ -59,7 +59,7 @@
     <Compile Include="DummyPlug.cs" />
     <Compile Include="Logger.cs" />
     <Compile Include="Model\Astroid.cs" />
-    <Compile Include="Model\Bot.cs" />
+    <Compile Include="Bot.cs" />
     <Compile Include="Model\Bullet.cs" />
     <Compile Include="ConnectionMapping.cs" />
     <Compile Include="ControlPanel.cs" />
@@ -73,12 +73,13 @@
     <Compile Include="GlobalSuppressions.cs" />
     <Compile Include="Model\Jet.cs" />
     <Compile Include="Model\Map.cs" />
+    <Compile Include="Model\Match.cs" />
     <Compile Include="Model\Player.cs" />
     <Compile Include="Model\Star.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Startup.cs" />
     <Compile Include="Model\Wall.cs" />
-    <Compile Include="ViewPort.cs" />
+    <Compile Include="Model\ViewPort.cs" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNet.Cors">
