+++ GameObjects/IUI.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Threading.Tasks;
+using GameObjects.Model;
 
 namespace GameObjects
 {
