+++ Planetes/Game.cs
@@ -5,6 +5,7 @@ using System.Drawing;
 using System.Drawing.Drawing2D;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using GameObjects.Model;
 
 namespace Planetes
 {
