+++ Planetes/HUD.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Linq;
 using System.Windows.Forms;
+using GameObjects.Model;
 
 namespace Planetes
 {
