+++ PlanetesWPF/HUD.xaml.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Media;
+using GameObjects.Model;
 
 namespace PlanetesWPF
 {
