+++ PlanetesWPF/UI.xaml.cs
@@ -6,6 +6,7 @@ using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Threading;
+using GameObjects.Model;
 
 namespace PlanetesWPF
 {
